Locate left and right peaks for heterodyne data
The original implementation didn't take into account that there must
be one and only one minor peak on either side of the center peak.

This update takes the a sorted list of height, index pairs and looks
for the next tallest peak that also satisfies the side-ness requirement.

This current implementation is hacky and needs to be updated further.