Construct single peak Lorentzian using optimal parameters
Previously, a Lorentzian was built using the inital guess parameters,
but no fitting function was actually called.

Now, the optimal parameters for the best fit Lorentzian is found.
Then, those parameters are used to build a Lorentzian. The result
is returned.

A later cell shows an example of this working on Channel B data.