WORKAROUND:  Fitting 3 peak Lorentzian then separating out
Since we only want to fit the side peaks for heterodyne data, it made
sense to attempt to fit a single peak Lorentzian to the heterodyne data
and "point" the curve fit to the desired peak at each call.
However, the resulting Lorentzians are hard to center and too wide.
Example is reproducible by uncommenting within `init_3lorentzian()`.

The workaround is to fit a 3 peak Lorentzian, extract the optimal
parameters from it, and create 3 single peak Lorentzian objects.
This process is different than the suggested "one-at-a-time" solution.

The quality of these results still needs to be quantified (chi-squared),
but visually are much better fits.