FIX - report mechanical frequency as location of center
Previously, this method returned the maximum peak value. However,
returning this value is incorrect for 2 reasons.

1. The units of max amplitude are m^2/Hz (or V^2/Hz). Such units
are not the correct units for measuring a frequency.

2. According to `fit_neg_sideband.dat` the (average) x-mode mechanical
frequency in the negative sideband is approximately 1.87 MHz.
This value is also supported graphically.

So the mechanical frequency is the frequency where the maximum amplitude
occurs, not the value of the maximum amplitude.