REFACTOR how experiment data directory is structured
On setup, raw and clean data directories are created.

The raw data directory only needs to create a child dir for the
UCL fit data. The split detection and heterodyne directories are
created by unzipping the packages while they are placed inside the
raw data directory.

For the clean data directory, subdirectories for heterodyne and
split detection are created. Then, blank template files are initialized.
The templates will serve as targets to write to during processing.