Refactor setup script
The top of the file depicts the desired filetree for organizing
raw and clean experiment data.

The bulk of the refactor is wrapping subdirectory creation and file
creation into separate functions.

The correct, public interface with this module is to call `build_file_tree()`
from `main.py`. This will create the file tree as specified above.