Remove reports and reporting from toolbox
At this stage, both UCL and CST data is processed.
The next deliverable are visualizations comparing time series between
these two data sources.

While reporting was useful, it was short sighted. The true comparison
is between the time series, not the exact values of the error coefficeints.
In a later phase, some of the concepts from reporting may come back to
supplement the visualizations.